+++ built/grammar/AbstractRuntimeCodeNode.js
@@ -57,18 +57,17 @@ class AbstractRuntimeCodeNode extends AbstractRuntimeNode_1.default {
     _getGrammarBackedCellArray() {
         const definition = this.getDefinition();
         const grammarProgram = definition.getProgram();
-        const parameters = definition.getNodeColumnTypes();
-        const expectedPattern = parameters.join(" ");
-        const parameterLength = parameters.length;
-        const lastParameterType = parameters[parameterLength - 1];
-        const lastParameterListType = lastParameterType && lastParameterType.endsWith("*") ? lastParameterType : undefined;
+        const columnTypes = definition.getNodeColumnTypes();
+        const expectedLinePattern = columnTypes.join(" ");
+        const numberOfColumns = columnTypes.length;
+        const lastColumnType = columnTypes[numberOfColumns - 1];
+        const lastColumnListType = lastColumnType && lastColumnType.endsWith("*") ? lastColumnType : undefined;
         const words = this.getWordsFrom(1);
-        const length = Math.max(words.length, parameterLength);
+        const length = Math.max(words.length, numberOfColumns);
         const checks = [];
+        // A for loop instead of map because "length" can be longer than words.length
         for (let wordIndex = 0; wordIndex < length; wordIndex++) {
-            const word = words[wordIndex];
-            const type = wordIndex >= parameterLength ? lastParameterListType : parameters[wordIndex];
-            checks[wordIndex] = new GrammarBackedCell_1.default(word, type, this, wordIndex, expectedPattern, grammarProgram);
+            checks[wordIndex] = new GrammarBackedCell_1.default(words[wordIndex], wordIndex >= numberOfColumns ? lastColumnListType : columnTypes[wordIndex], this, wordIndex, expectedLinePattern, grammarProgram);
         }
         return checks;
     }