+++ src/grammar/GrammarConstants.ts
@@ -38,6 +38,7 @@ GrammarConstants.compiler.closeChildren = "@closeChildren"
 // developing
 GrammarConstants.description = "@description"
 GrammarConstants.frequency = "@frequency"
+GrammarConstants.highlightScope = "@highlightScope"
 
 // errors
 GrammarConstants.invalidKeywordError = "invalidKeywordError"